Update capture
i took out the need for a metadata file. you need to have two input folders called fake_img and real_img. these folders can contain videos(.mp4) and image(.jpg). and you need to have an empty folder called dataset, inside of this you need two empty folders called real and fake. this is where the dataset of faces are stored. i also added a check to make sure that we only crop an image if a face is detected and we only add the image to the dataset if the crop image is not "empty"